Move more operator-related items (#638)
